Support Vernissage 2.3 as well
The Vernissage SDK did not change since version 2.2 so we can, for the
first time, support two major Vernissage versions.

We do that by trying multiple registry keys (required due to name changing
of the company).

We now also read the "RecentVersion" registry value as multiple versions
of Vernissage can now be installed on a machine.

If multiple Vernissage versions are installed the last installed version will be
used.