Fix encoding of metadata/text waves for IP7
In 3836b370 (Work around regression suite failures for Igor 7, 2015-07-22)
the encoding of the created text waves was set to Windows 1252 (3 in IP7)
to fix some test failures.

This was of course the wrong approach as we should create text waves with
UTF8 encoding instead of fixing them up later on.