Tests: Shortend directory names in the new version
With the new support for long object names, the directories on the disc
are now longer as well, as they are derived from IP datafolders.

But then comparing old and new folders breaks as the old ones are
shorter.

By replacing known longer ones with shorter ones we can make the check
work.