MFR_GetBrickletData: Fix resampling/pixelation errors
When loading bricklets which have too few points for the resampling or
pixelation, we currently bug out or skip the bricklet depending wether
it is 1D or 2D.

While this can be manually worked around, this currently breaks the MFR
GUI as there is no fallback for failed resampling/pixelation.

We now check that we have enough points before trying the pixelation.
And we also only set the resampled dimension scales when resampling
worked, and also use the original wave when resampling failed.

Reported-By: Nils Krane <Nils.Krane@empa.ch>